#17 Move the projects to .net core 2.1
