+++ src/Easify.AspNetCore/ExceptionHandling/ExceptionHandlingExtensions.cs
@@ -1,0 +1,40 @@
+// This software is part of the Easify framework
+// Copyright (C) 2019 Intermediate Capital Group
+// 
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Affero General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+// 
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU Affero General Public License for more details.
+// 
+// You should have received a copy of the GNU Affero General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+using System;
+using Easify.ExceptionHandling.ErrorBuilder;
+using Easify.ExceptionHandling.ErrorBuilder.Fluent;
+using Microsoft.AspNetCore.Builder;
+using RestEase;
+
+namespace Easify.AspNetCore.ExceptionHandling
+{
+    public static class ExceptionHandlingExtensions
+    {
+        public static IApplicationBuilder UseGlobalExceptionHandler(this IApplicationBuilder app)
+        {
+            if (app == null) throw new ArgumentNullException(nameof(app));
+            app.UseMiddleware<ErrorHandlingMiddleware>();
+            return app;
+        }
+
+        public static IProvideErrorBuilder<ApiException> UseBuilderForApi(
+            this ISetErrorBuilder<ApiException> builderProvider)
+        {
+            return builderProvider.Use(new ErrorBuilderForApiException());
+        }
+    }
+}
