+++ src/Easify.Logging.SeriLog.Loggly/Easify.Logging.SeriLog.Loggly.csproj
@@ -13,7 +13,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="2.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="2.1.1" />
     <PackageReference Include="Serilog.Sinks.Loggly" Version="5.1.1" />
   </ItemGroup>
