+++ src/Easify.Logging.SeriLog.Seq/Easify.Logging.SeriLog.Seq.csproj
@@ -13,7 +13,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="2.0.0" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="2.1.1" />
     <PackageReference Include="Serilog.Sinks.Seq" Version="3.4.0" />
   </ItemGroup>
