+++ src/Easify.RestEase/Easify.RestEase.csproj
@@ -11,8 +11,8 @@
     <Version>1.0.0</Version>
   </PropertyGroup>
   <ItemGroup>
-      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.0.0" />
-      <PackageReference Include="Microsoft.Extensions.Options" Version="2.0.0" />
+      <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="2.1.1" />
+      <PackageReference Include="Microsoft.Extensions.Options" Version="2.1.1" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\Easify.RestEase.Client\Easify.RestEase.Client.csproj" />
