+++ src/Easify.Testing.Integration/Easify.Testing.Integration.csproj
@@ -9,10 +9,10 @@
         <Version>1.0.0</Version>
     </PropertyGroup>
     <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.Server.Kestrel" Version="2.0.1"/>
-        <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="2.0.1"/>
-        <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.0.0"/>
-        <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.0.0"/>
+        <PackageReference Include="Microsoft.AspNetCore.Server.Kestrel" Version="2.1.1"/>
+        <PackageReference Include="Microsoft.AspNetCore.TestHost" Version="2.1.1"/>
+        <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.1.1"/>
+        <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="2.1.1"/>
     </ItemGroup>
 
     <ItemGroup>
