+++ src/Easify.Testing/Easify.Testing.csproj
@@ -11,11 +11,12 @@
     
     <ItemGroup>
         <PackageReference Include="xunit.abstractions" Version="2.0.1" />
-        <PackageReference Include="AutoFixture" Version="4.5.0" />
-        <PackageReference Include="AutoFixture.AutoNSubstitute" Version="4.5.0" />
-        <PackageReference Include="AutoFixture.Xunit2" Version="4.5.0" />
+        <PackageReference Include="Castle.Core" Version="4.4.1" />
+        <PackageReference Include="AutoFixture" Version="4.11.0" />
+        <PackageReference Include="AutoFixture.AutoNSubstitute" Version="4.11.0" />
+        <PackageReference Include="AutoFixture.Xunit2" Version="4.11.0" />
         <PackageReference Include="FluentAssertions" Version="5.4.2" />
-        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="2.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="2.1.1" />
         <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
     </ItemGroup>
